pyproject.toml
[tool.poetry]
name = "mcp-test"
version = "0.1.0"
description = "MCP Test Project"
authors = ["Olivier J.N. Bertrand <bolirev@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-storage = "^3.1.0"
python-dotenv = "^1.0.1"
langchain-openai = "*"
langgraph = "*"
mcp = "^1.9.2"
langchain-mcp-adapters = "^0.1.7"
jupyter = "^1.1.1"

[tool.poetry.dev-dependencies]
#jupyter = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
